+++ lib/collection/src/shards/replica_set/snapshots.rs
@@ -1,4 +1,5 @@
 use std::path::Path;
+use std::sync::Arc;
 
 use common::fs::{safe_delete_with_suffix, sync_parent_dir_async};
 use common::save_on_disk::SaveOnDisk;
@@ -84,6 +85,44 @@ impl ShardReplicaSet {
         &self,
     ) -> CollectionResult<tokio::sync::OwnedRwLockWriteGuard<()>> {
         self.partial_snapshot_meta.try_take_recovery_lock()
+    }
+
+    /// Exclusive access to full snapshot recovery of this shard, waiting for one in
+    /// progress to finish. Waiters are served in arrival order.
+    ///
+    /// Must be held across the whole recovery - clear, download and restore - each of
+    /// which is destructive on its own. While held, no other recovery of this shard can
+    /// make progress, so the recovery that reports success is the last one to touch the
+    /// shard and cannot be rolled back by one that is still in flight.
+    ///
+    /// Callers go through [`Collection::start_shard_recovery`], which folds this into the
+    /// [`ShardRecoveryGuard`] rather than leaving it to be held separately.
+    ///
+    /// [`Collection::start_shard_recovery`]: crate::collection::Collection::start_shard_recovery
+    /// [`ShardRecoveryGuard`]: crate::shards::shard_holder::recovery_guard::ShardRecoveryGuard
+    pub(crate) async fn take_snapshot_recovery_lock(&self) -> tokio::sync::OwnedMutexGuard<()> {
+        if let Ok(recovery_lock) = Arc::clone(&self.snapshot_recovery_lock).try_lock_owned() {
+            return recovery_lock;
+        }
+
+        log::warn!(
+            "Snapshot recovery of shard {}:{} is waiting for a recovery of the same shard \
+             that is already in progress",
+            self.collection_id,
+            self.shard_id,
+        );
+
+        let waiting_since = std::time::Instant::now();
+        let recovery_lock = Arc::clone(&self.snapshot_recovery_lock).lock_owned().await;
+
+        log::info!(
+            "Snapshot recovery of shard {}:{} waited {:.1}s for the previous recovery to finish",
+            self.collection_id,
+            self.shard_id,
+            waiting_since.elapsed().as_secs_f32(),
+        );
+
+        recovery_lock
     }
 
     pub fn restore_snapshot(
@@ -450,187 +489,4 @@ async fn wait_restore_local_replica_before_flag_hook_for_test(shard_flag: &Path)
 }
 
 #[cfg(test)]
-mod tests {
-    use std::collections::HashSet;
-    use std::num::NonZeroU32;
-    use std::sync::Arc;
-
-    use common::budget::ResourceBudget;
-    use common::save_on_disk::SaveOnDisk;
-    use segment::types::Distance;
-    use tempfile::{Builder, TempDir};
-    use tokio::runtime::Handle;
-    use tokio::sync::{RwLock, oneshot};
-
-    use super::*;
-    use crate::collection::payload_index_schema::PayloadIndexSchema;
-    use crate::common::adaptive_handle::AdaptiveSearchHandle;
-    use crate::config::{CollectionConfigInternal, CollectionParams, WalConfig};
-    use crate::operations::shared_storage_config::SharedStorageConfig;
-    use crate::operations::types::VectorsConfig;
-    use crate::operations::vector_params_builder::VectorParamsBuilder;
-    use crate::optimizers_builder::OptimizersConfig;
-    use crate::shards::channel_service::ChannelService;
-    use crate::shards::replica_set::replica_set_state::ReplicaState;
-    use crate::shards::replica_set::{AbortShardTransfer, ChangePeerFromState};
-    use crate::shards::shard::ShardId;
-
-    const TEST_COLLECTION_ID: &str = "test_collection";
-    const TEST_TARGET_SHARD_ID: ShardId = 1;
-    const TEST_SOURCE_SHARD_ID: ShardId = 2;
-    const TEST_PEER_ID: PeerId = 1;
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_cancel_snapshot_recovery_before_initializing_flag_does_not_mark_dirty() {
-        let target_collection_dir = Builder::new()
-            .prefix("snapshot-recovery-cancel-target")
-            .tempdir()
-            .unwrap();
-        let source_collection_dir = Builder::new()
-            .prefix("snapshot-recovery-cancel-source")
-            .tempdir()
-            .unwrap();
-
-        let target_replica_set =
-            new_shard_replica_set(&target_collection_dir, TEST_TARGET_SHARD_ID).await;
-        let source_replica_set =
-            new_shard_replica_set(&source_collection_dir, TEST_SOURCE_SHARD_ID).await;
-
-        let source_shard_path = source_collection_dir
-            .path()
-            .join(TEST_SOURCE_SHARD_ID.to_string());
-        assert!(
-            LocalShard::check_data(&source_shard_path),
-            "test fixture must create a valid source local shard"
-        );
-
-        let shard_flag =
-            shard_initializing_flag_path(target_collection_dir.path(), TEST_TARGET_SHARD_ID);
-        let (reached_tx, reached_rx) = oneshot::channel();
-        let (release_tx, release_rx) = oneshot::channel();
-        install_restore_local_replica_before_flag_hook(shard_flag.clone(), reached_tx, release_rx);
-
-        let cancel = cancel::CancellationToken::new();
-        let restore = target_replica_set.restore_local_replica_from(
-            &source_shard_path,
-            RecoveryType::Full,
-            target_collection_dir.path(),
-            cancel.clone(),
-        );
-        tokio::pin!(restore);
-
-        tokio::select! {
-            _ = reached_rx => {}
-            result = &mut restore => {
-                panic!("snapshot recovery completed before the cancellation window: {result:?}");
-            }
-        }
-
-        cancel.cancel();
-        let _ = release_tx.send(());
-
-        let result = restore.await;
-        assert!(
-            matches!(result, Err(CollectionError::Cancelled { .. })),
-            "recovery should observe cancellation before creating the initializing flag, got {result:?}"
-        );
-        assert!(
-            !shard_flag.exists(),
-            "cancellation before destructive restore starts must not leave a false dirty marker"
-        );
-        assert!(
-            !target_replica_set.is_dummy().await,
-            "cancellation before destructive restore starts must keep the old local shard installed"
-        );
-
-        let local = target_replica_set.local.read().await;
-        assert!(
-            matches!(local.as_ref(), Some(Shard::Local(_))),
-            "the old local shard should remain installed"
-        );
-        drop(local);
-
-        target_replica_set.stop_gracefully().await;
-        source_replica_set.stop_gracefully().await;
-    }
-
-    fn install_restore_local_replica_before_flag_hook(
-        shard_flag: std::path::PathBuf,
-        reached: oneshot::Sender<()>,
-        release: oneshot::Receiver<()>,
-    ) {
-        let mut hook = RESTORE_LOCAL_REPLICA_BEFORE_FLAG_HOOK.lock().unwrap();
-        assert!(
-            hook.is_none(),
-            "restore-local-replica test hook is already installed"
-        );
-        *hook = Some((shard_flag, reached, release));
-    }
-
-    async fn new_shard_replica_set(collection_dir: &TempDir, shard_id: ShardId) -> ShardReplicaSet {
-        let update_runtime = Handle::current();
-        let search_runtime = AdaptiveSearchHandle::current_for_tests();
-
-        let wal_config = WalConfig {
-            wal_capacity_mb: 1,
-            wal_segments_ahead: 0,
-            wal_retain_closed: 1,
-        };
-
-        let collection_params = CollectionParams {
-            vectors: VectorsConfig::Single(VectorParamsBuilder::new(4, Distance::Dot).build()),
-            shard_number: NonZeroU32::new(1).unwrap(),
-            replication_factor: NonZeroU32::new(1).unwrap(),
-            write_consistency_factor: NonZeroU32::new(1).unwrap(),
-            ..CollectionParams::empty()
-        };
-
-        let optimizers_config = OptimizersConfig::fixture();
-        let config = CollectionConfigInternal {
-            params: collection_params,
-            optimizer_config: optimizers_config.clone(),
-            wal_config,
-            hnsw_config: Default::default(),
-            quantization_config: None,
-            strict_mode_config: None,
-            uuid: None,
-            metadata: None,
-        };
-
-        let payload_index_schema_file = collection_dir.path().join("payload-schema.json");
-        let payload_index_schema: Arc<SaveOnDisk<PayloadIndexSchema>> =
-            Arc::new(SaveOnDisk::load_or_init_default(payload_index_schema_file).unwrap());
-        let shared_config = Arc::new(RwLock::new(config));
-
-        ShardReplicaSet::build(
-            shard_id,
-            None,
-            TEST_COLLECTION_ID.to_string(),
-            TEST_PEER_ID,
-            true,
-            HashSet::new(),
-            dummy_on_replica_failure(),
-            dummy_abort_shard_transfer(),
-            collection_dir.path(),
-            shared_config,
-            optimizers_config.clone(),
-            Arc::new(SharedStorageConfig::default()),
-            payload_index_schema,
-            ChannelService::default(),
-            update_runtime,
-            search_runtime,
-            ResourceBudget::default(),
-            Some(ReplicaState::Active),
-        )
-        .await
-        .unwrap()
-    }
-
-    fn dummy_on_replica_failure() -> ChangePeerFromState {
-        Arc::new(move |_peer_id, _shard_id, _from_state| {})
-    }
-
-    fn dummy_abort_shard_transfer() -> AbortShardTransfer {
-        Arc::new(|_shard_transfer, _reason| {})
-    }
-}
+mod tests;
